get one, update and like functionality

+++ backend/src/sitios/sitios.service.ts
@@ -1,5 +1,5 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
-import { CrearSitioDTO } from './sitio.dto';
+import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
+import { UpdateSitioDTO, CreateSitioDTO } from './sitio.dto';
 import { SitioEntity } from './sitio.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
@@ -11,11 +11,19 @@ export class SitiosService {
     private sitioRepository: Repository<SitioEntity>,
   ) { }
 
-  async retrieve() {
+  async find() {
     return await this.sitioRepository.find();
   }
 
-  async create(sitioDTO: CrearSitioDTO) {
+  async findOne(id: number) {
+    const sitio = this.sitioRepository.findOneBy({ id });
+    if (!sitio) {
+      throw new NotFoundException(`Site with ID ${id} not found.`);
+    }
+    return sitio;
+  }
+
+  async create(sitioDTO: CreateSitioDTO) {
     try {
       return await this.sitioRepository.save(sitioDTO);
     } catch (err) {
@@ -23,7 +31,50 @@ export class SitiosService {
     }
   }
 
+  async like(id: number) {
+
+    const sitio = await this.sitioRepository.findOneBy({ id });
+
+    if (!sitio) {
+      throw new NotFoundException(`Site with ID ${id} not found.`);
+    }
+
+    try {
+      Object.assign(sitio, {
+        ...sitio,
+        likes: sitio.likes + 1,
+      });
+
+      return await this.sitioRepository.save(sitio);
+    } catch (err) {
+      throw new InternalServerErrorException("Could't save site to the database");
+    }
+  }
+
+  async update(id: number, sitioDTO: UpdateSitioDTO) {
+
+    const sitio = await this.sitioRepository.findOneBy({ id });
+
+    if (!sitio) {
+      throw new NotFoundException(`Site with ID ${id} not found.`);
+    }
+
+    try {
+      Object.assign(sitio, sitioDTO);
+
+      return await this.sitioRepository.save(sitio);
+    } catch (err) {
+      throw new InternalServerErrorException("Could't save site to the database");
+    }
+  }
+
   async delete(id: number) {
+    const sitio = await this.sitioRepository.findOneBy({ id });
+
+    if (!sitio) {
+      throw new NotFoundException(`Site with ID ${id} not found.`);
+    }
+
     return await this.sitioRepository.delete(id);
   }
 }
